backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
psycopg2-binary
requests
beautifulsoup4
spacy
textblob
python-dotenv
